+++ ios/appmetrica_push_plugin/Sources/appmetrica_push_plugin/AMPFAppMetricaPushImplementation.m
@@ -56,6 +56,10 @@
     }
 }
 
+- (void)enableLoggerWithError:(FlutterError **)error {
+    // do nothing
+}
+
 - (void)setUserInfo:(NSDictionary *)userInfo
 {
     _userInfo = userInfo;