+++ lib/src/appmetrica_push_api_pigeon.dart
@@ -164,6 +164,28 @@ class AppMetricaPushPigeon {
     }
   }
 
+  Future<void> enableLogger() async {
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.appmetrica_push_plugin.AppMetricaPushPigeon.enableLogger', codec,
+        binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList =
+        await channel.send(null) as List<Object?>?;
+    if (replyList == null) {
+      throw PlatformException(
+        code: 'channel-error',
+        message: 'Unable to establish connection on channel.',
+      );
+    } else if (replyList.length > 1) {
+      throw PlatformException(
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
+      );
+    } else {
+      return;
+    }
+  }
+
   Future<Map<String?, String?>> getTokens() async {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.appmetrica_push_plugin.AppMetricaPushPigeon.getTokens', codec,